Add rule for prop types import
