+++ package.json
@@ -10,14 +10,16 @@
     "lint"
   ],
   "devDependencies": {
-    "eslint": "~2.8.0",
-    "babel-preset-es2015": "~6.22.0"
+    "babel-preset-es2015": "^6.22.0",
+    "eslint": "~2.8.0"
   },
   "scripts": {
     "postinstall": "babel src --out-dir lib"
   },
   "babel": {
-    "presets": ["es2015"]
+    "presets": [
+      "es2015"
+    ]
   },
   "engines": {
     "node": ">=6.2"
